Add 4.21 - blogilistan laajennus, step9

+++ controllers/blogs.js
@@ -44,9 +44,19 @@ blogsRouter.post('/', async (request, response, next) => {
   }
 });
 
-blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id);
-  response.status(204).end();
+blogsRouter.delete('/:id', async (request, response, next) => {
+  try {
+    const token = request.token;
+    const blog = await Blog.findById(request.params.id);
+    const decodedToken = jwt.verify(token, process.env.SECRET);
+
+    if (blog.user.toString() === decodedToken.id.toString()) {
+      await Blog.findByIdAndRemove(request.params.id);
+      response.status(204).end();
+    }
+  } catch (err) {
+    next(err);
+  }
 });
 
 blogsRouter.put('/:id', async (request, response) => {
